drivers: prevent unnecessary, double modifications
The double modification results in a "-Wsequence-point" warning from the
C compiler. Making two modifications on the same line of code isn't
really a sensible concept anyway, and was clearly not intended here.

The unnecessary modifications will probably be optimized out, but it's
cleaner to not code them to begin with.