rewrite ad1939 driver
This is a *very minimal* SPI driver for the ad1939. It only initializes
the hardware to the exact settings we use in class with the Audio Mini.

This update was needed because `spi_busnum_to_master` has been
deprecated, so the way we used to handle the ad1939 spi device is not
possible.

The ad1939 must be a slave under spi0 in the device tree.